refactor: Change project structure to have a component structure
