+++ gulpfile.js
@@ -90,7 +90,7 @@ var tsProject = plugins.typescript.createProject('tsconfig.json', {
 });
 
 function ts() {
-	var tsResult = gulp.src('src/**/*.ts')
+	var tsResult = gulp.src('src/assets/app/**/*')
 		.pipe(plugins.if(argv.dev, plugins.sourcemaps.init()))
 		.pipe(plugins.typescript(tsProject));
